requirements.txt
streamlit==1.35.0
pandas==2.2.2
numpy==1.26.4
scikit-learn==1.5.0
Pillow==10.3.0
streamlit-extras==0.4.0
